刘恒 exam增加QuestionReq 试题动态查询 课程id改成Long

+++ src/main/kotlin/com/huayei/domain/exam/testQuestions/request/QuestionReq.kt
@@ -14,6 +14,6 @@ import com.huayei.base.PageReq
  * @Since 1.0
  **/
 data class QuestionReq (
-    var courseId: Int? = null,
+    var courseId: Long? = null,
     var questionType: String? = null
 ) : PageReq()
